+++ R/get_predictions_gam2.R
@@ -7,6 +7,10 @@ get_predictions_Gam <- function(model, fitfram, ci.lvl, linv, value_adjustment, 
   else
     ci <- 0.975
 
+  # degrees of freedom
+  dof <- .get_df(model)
+  tcrit <- stats::qt(ci, df = dof)
+
   prdat <-
     stats::predict(
       model,
@@ -36,8 +40,8 @@ get_predictions_Gam <- function(model, fitfram, ci.lvl, linv, value_adjustment, 
       fitfram <- se.pred$prediction_data
 
       # calculate CI
-      fitfram$conf.low <- linv(as.vector(prdat) - stats::qnorm(ci) * se.fit)
-      fitfram$conf.high <- linv(as.vector(prdat) + stats::qnorm(ci) * se.fit)
+      fitfram$conf.low <- linv(as.vector(prdat) - tcrit * se.fit)
+      fitfram$conf.high <- linv(as.vector(prdat) + tcrit * se.fit)
 
       # copy standard errors
       attr(fitfram, "std.error") <- se.fit