+++ R/get_predictions_mclogit.R
@@ -8,6 +8,10 @@ get_predictions_mclogit <- function(model, fitfram, ci.lvl, model_class, value_a
   else
     ci <- 0.975
 
+  # degrees of freedom
+  dof <- .get_df(model)
+  tcrit <- stats::qt(ci, df = dof)
+
   # add response to new data
   resp <- insight::find_response(model, combine = FALSE)
   cn <- c(colnames(fitfram), resp)
@@ -16,14 +20,13 @@ get_predictions_mclogit <- function(model, fitfram, ci.lvl, model_class, value_a
   }
   colnames(fitfram) <- cn
 
-  prdat <-
-    stats::predict(
-      model,
-      newdata = fitfram,
-      type = "response",
-      se.fit = se,
-      ...
-    )
+  prdat <- stats::predict(
+    model,
+    newdata = fitfram,
+    type = "response",
+    se.fit = se,
+    ...
+  )
 
   # did user request standard errors? if yes, compute CI
   if (!is.null(vcov.fun)) {
@@ -52,8 +55,8 @@ get_predictions_mclogit <- function(model, fitfram, ci.lvl, model_class, value_a
       fitfram <- se.pred$prediction_data
 
       # CI
-      fitfram$conf.low <- fitfram$predicted - stats::qnorm(ci) * se.fit
-      fitfram$conf.high <- fitfram$predicted + stats::qnorm(ci) * se.fit
+      fitfram$conf.low <- fitfram$predicted - tcrit * se.fit
+      fitfram$conf.high <- fitfram$predicted + tcrit * se.fit
 
       # copy standard errors
       attr(fitfram, "std.error") <- se.fit
@@ -68,8 +71,8 @@ get_predictions_mclogit <- function(model, fitfram, ci.lvl, model_class, value_a
     fitfram$predicted <- prdat$fit
 
     # calculate CI
-    fitfram$conf.low <- prdat$fit - stats::qnorm(ci) * prdat$se.fit
-    fitfram$conf.high <- prdat$fit + stats::qnorm(ci) * prdat$se.fit
+    fitfram$conf.low <- prdat$fit - tcrit * prdat$se.fit
+    fitfram$conf.high <- prdat$fit + tcrit * prdat$se.fit
 
     # copy standard errors
     attr(fitfram, "std.error") <- prdat$se.fit