+++ R/get_predictions_tobit.R
@@ -8,14 +8,17 @@ get_predictions_tobit <- function(model, fitfram, ci.lvl, linv, ...) {
   else
     ci <- 0.975
 
-  prdat <-
-    stats::predict(
-      model,
-      newdata = fitfram,
-      type = "lp",
-      se.fit = se,
-      ...
-    )
+  # degrees of freedom
+  dof <- .get_df(model)
+  tcrit <- stats::qt(ci, df = dof)
+
+  prdat <- stats::predict(
+    model,
+    newdata = fitfram,
+    type = "lp",
+    se.fit = se,
+    ...
+  )
 
   # did user request standard errors? if yes, compute CI
   if (se) {
@@ -23,8 +26,8 @@ get_predictions_tobit <- function(model, fitfram, ci.lvl, linv, ...) {
     fitfram$predicted <- linv(prdat$fit)
 
     # calculate CI
-    fitfram$conf.low <- linv(prdat$fit - stats::qnorm(ci) * prdat$se.fit)
-    fitfram$conf.high <- linv(prdat$fit + stats::qnorm(ci) * prdat$se.fit)
+    fitfram$conf.low <- linv(prdat$fit - tcrit * prdat$se.fit)
+    fitfram$conf.high <- linv(prdat$fit + tcrit * prdat$se.fit)
 
     # copy standard errors
     attr(fitfram, "std.error") <- prdat$se.fit